fix(web): preserve sub-millisecond precision in p95 / p99 / wait
The initial integer-microsecond → integer-millisecond fix in c1567be
collapsed any percentile under 1 ms to zero. A pool whose true p95
was 420 µs reported `query_p95_ms = 0` and the dashboard rendered
a "0 ms" tile that contradicted the log line `query_ms p95 = 0.42`.

Switch the affected DTO fields to `f64` (`query_p95_ms`,
`query_p99_ms`, `transactions_p95_ms`, `transactions_p99_ms`,
`wait_avg_ms`, `wait_p95_ms`) and divide as floating-point. Front-
end formatter now renders sub-1 ms with two decimals and 1-9 ms
with one decimal, so a 0.42-ms p95 shows up as `0.42ms` and a
6.6-ms p95 shows up as `6.6ms` instead of being rounded to "1ms".

Threshold comparisons (warn at 100 ms, danger at 500 ms) work the
same against `f64` as they did against `u64`. JSON serialisation
emits the value as a JS number which TypeScript already typed.

+++ src/web/routes/dto.rs
@@ -92,13 +92,18 @@ pub(crate) struct PoolDto {
 
     pub max_active_age_ms: u64,
 
-    pub query_p95_ms: u64,
-    pub query_p99_ms: u64,
-    pub transactions_p95_ms: u64,
-    pub transactions_p99_ms: u64,
+    /// Latency percentiles in milliseconds. Stored as `f64` rather than
+    /// `u64` so sub-millisecond values survive: a workload whose true
+    /// p95 is 420 µs would otherwise integer-divide to `0 ms` in the
+    /// DTO and render as a bogus zero on the dashboard while the log
+    /// line correctly showed `query_ms p95 = 0.42`.
+    pub query_p95_ms: f64,
+    pub query_p99_ms: f64,
+    pub transactions_p95_ms: f64,
+    pub transactions_p99_ms: f64,
 
-    pub wait_avg_ms: u64,
-    pub wait_p95_ms: u64,
+    pub wait_avg_ms: f64,
+    pub wait_p95_ms: f64,
 
     pub queries_total: u64,
     pub transactions_total: u64,
